Created Basic Enemy & IAttackable Interface
Created the BaseEnemy which currently just walks towards a given actor. Also created the IAttackable interface which will be given to anything in the game that can be attacked. They are forced to have an enum called Team which represents which team they are on so that all attackable objects are able to tell if another object is on the same of a different team to them, therefore letting them know if they should chase/attack them.